chore: several renames and improvements
