+++ packages/components/modal/test/ModalConfirm.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { axe } from '@/scripts/test/axeHelper';
 
-import { ModalConfirm } from './ModalConfirm';
+import { ModalConfirm } from '../src/ModalConfirm/ModalConfirm';
 import '@contentful/forma-36-fcss/dist/styles.css';
 
 jest.mock(
